Fix bug in tiff_timestamp_checker epoch assignment
Previously we were assigning epochs based on chunk_length_daq, which is
not correct as some of these triggers are discarded and aren't real
frames. This issue doesn't manifest apart from in crashed sessions which
can have lots of extra frames, causing the epoch to be wrongly assigned
to the previous one.